requirements.txt
# Neural Watch - Phase 1 Requirements
# Data Ingestion & Quality Setup

# Backend Framework
fastapi
uvicorn[standard]
python-multipart
pydantic

# uvicorn dependency
click

# Data Processing
pandas
numpy
pyarrow

# Phase 2: Statistical Analysis
scipy
scikit-learn

# Visualization (Phase 2)
matplotlib
seaborn
plotly

# Environment & Configuration
python-dotenv

# Frontend
streamlit

# HTTP Requests
requests

# Testing
pytest
pytest-cov
pytest-asyncio

# Logging & Utilities
colorlog

# Future Phases (commented out for now)
# Phase 2-3: Statistical Tests
# scipy==1.11.4
# scikit-learn==1.4.0

# Phase 4: LLM Integration
# langchain==0.1.0
# openai==1.10.0
# anthropic==0.8.0

# Phase 5: Alerts
# slack-sdk==3.26.2

# Streaming (Optional)
# kafka-python==2.0.2

# Visualization
# matplotlib==3.8.2
# seaborn==0.13.1
# plotly==5.18.0